// src/types.ts
// Generated by https://quicktype.io
//
// To change quicktype's target language, run command:
//
//   "Set quicktype target language"

export interface NewsElement {
    sophoraId:      string;
    externalId:     string;
    title:          string;
    teaserImage?:   TeaserImage;
    date:           string;
    tracking:       { [key: string]: string }[];
    tags:           Tag[];
    updateCheckUrl: string;
    regionId?:      number;
    details?:       string;
    detailsweb?:    string;
    shareURL?:      string;
    topline?:       string;
    firstSentence?: string;
    geotags?:       Tag[];
    ressort?:       Ressort;
    type:           NewsType;
    regionIds?:     number[];
    brandingImage?: BrandingImage;
    streams?:       { [key: string]: string };
    copyright?:     string;
    alttext?:       string;
}

export interface BrandingImage {
    title:     Alttext;
    copyright: Copyright;
    alttext:   Alttext;
    type:      BrandingImageType;
    videowebm: Imageurl;
    videowebs: Imageurl;
}

export enum Alttext {
    LogoBR = "Logo BR",
    LogoHR = "Logo HR",
    LogoMDR = "Logo MDR",
    LogoNDR = "Logo NDR",
    LogoRB = "Logo RB",
    LogoRbb = "Logo rbb",
    LogoSR = "Logo SR",
    LogoSWR = "Logo SWR",
    LogoSportschau = "Logo Sportschau",
    LogoWDR = "Logo WDR",
}

export enum Copyright {
    TagesschauDeARDBR = "tagesschau.de, ARD, BR",
    TagesschauDeARDHR = "tagesschau.de, ARD, HR",
    TagesschauDeARDMDR = "tagesschau.de, ARD, MDR",
    TagesschauDeARDNDR = "tagesschau.de, ARD, NDR",
    TagesschauDeARDRB = "tagesschau.de, ARD, RB",
    TagesschauDeARDRbb = "tagesschau.de, ARD, rbb",
    TagesschauDeARDSR = "tagesschau.de, ARD, SR",
    TagesschauDeARDSWR = "tagesschau.de, ARD, SWR",
    TagesschauDeARDSportschau = "tagesschau.de, ARD, Sportschau",
    TagesschauDeARDWDR = "tagesschau.de, ARD, WDR",
}

export enum BrandingImageType {
    Image = "image",
}

export interface Imageurl {
    imageurl: string;
}

export interface Tag {
    tag: string;
}

export enum Ressort {
    Ausland = "ausland",
    Inland = "inland",
    Investigativ = "investigativ",
    Sport = "sport",
    Wirtschaft = "wirtschaft",
}

export interface TeaserImage {
    title:                 string;
    copyright?:            string;
    alttext:               string;
    type:                  BrandingImageType;
    videowebl:             Imageurl;
    klein1x1?:             Imageurl;
    portraetgrossplus8x9?: Imageurl;
    videoweb1x1l?:         Imageurl;
    videowebm:             Imageurl;
    videowebs:             Imageurl;
    portraetgross8x9?:     Imageurl;
    mittelgross1x1?:       Imageurl;
}

export enum NewsType {
    Story = "story",
    Video = "video",
    Webview = "webview",
}